% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourvalplot.R
\name{fourvalplot}
\alias{fourvalplot}
\title{Plot spider chart for four higher order personal values}
\usage{
fourvalplot(df, valueMap, corrected = TRUE, standerr = TRUE, na.rm = TRUE)
}
\arguments{
\item{df}{A dataframe containing computed values for the four higher order personal values.}

\item{valueMap}{A named vector where names are the four higher order personal values:
Openness to Change ("OPEN"), Self-enhancement ("SELFENH"), Conservation ("CONS"),
Self-transcendence ("SELFTRANS"). Vector values correspond to the column names in the dataframe.}

\item{corrected}{Logical. Indicates if data are ipsatized (range -4 to +4); if FALSE, data are raw scores (range 1 to 5).}

\item{standerr}{Logical. When TRUE, plots mean ± SE lines; when FALSE, plots only the mean.}

\item{na.rm}{Logical. When TRUE, NAs are ignored in calculations; when FALSE, NAs affect calculations.}
}
\value{
A spider chart visualizing the four higher order values.
}
\description{
This function generates a spider chart based on four higher order Basic Human Values:
Openness to Change, Self-Enhancement, Conservation, and Self-Transcendence.
This visualization helps in understanding the profile of higher order values for individuals or groups.
}
\examples{
df <- data.frame(
  open_comp = runif(10, 1, 5), selfenh_comp = runif(10, 1, 5),
  cons_comp = runif(10, 1, 5), selftrans_comp = runif(10, 1, 5)
)
valueMap <- c(OPEN = "open_comp", SELFENH = "selfenh_comp",
              CONS = "cons_comp", SELFTRANS = "selftrans_comp")
fourvalplot(df, valueMap, corrected = TRUE, standerr = TRUE)
}
