% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenvalplot.R
\name{tenvalplot}
\alias{tenvalplot}
\title{Plot spider chart for ten specific personal values}
\usage{
tenvalplot(df, valueMap, corrected = TRUE, standerr = TRUE, na.rm = TRUE)
}
\arguments{
\item{df}{A dataframe containing the computed values for the ten specific personal values.}

\item{valueMap}{A named vector where names correspond to the ten personal values:
Conformity ("CO"), Tradition ("TR"), Benevolence ("BE"), Achievement ("AC"),
Power ("PO"), Security ("SE"), Stimulation ("ST"), Self-direction ("SD"),
Universalism ("UN"), Hedonism ("HE"). Vector values correspond to the column names
in the dataframe.}

\item{corrected}{Logical. When TRUE, data are already ipsatized (range -4 to +4);
when FALSE, data are raw scores (range 1 to 5). Defaults to TRUE.}

\item{standerr}{Logical. When TRUE, plots mean-SE and mean+SE lines; when FALSE,
plots only the mean. Default is TRUE.}

\item{na.rm}{Logical. When TRUE, NAs are ignored in calculations; when FALSE, NAs
are preserved and will affect calculations. Default is TRUE.}
}
\description{
This function generates a spider chart based on the Basic Human Values.
It requires exactly ten value categories corresponding to the 10 predefined
personal values. This visualization helps in understanding the individual
or average profile of personal values.
}
\examples{
df <- data.frame(
  conf_comp = runif(10, 1, 5), trad_comp = runif(10, 1, 5),
  bene_comp = runif(10, 1, 5), achie_comp = runif(10, 1, 5),
  power_comp = runif(10, 1, 5), sec_comp = runif(10, 1, 5),
  stim_comp = runif(10, 1, 5), selfdir_comp = runif(10, 1, 5),
  univ_comp = runif(10, 1, 5), hedo_comp = runif(10, 1, 5)
)
valueMap <- c(CO = "conf_comp", TR = "trad_comp",
              BE = "bene_comp", AC = "achie_comp",
              PO = "power_comp", SE = "sec_comp",
              ST = "stim_comp", SD = "selfdir_comp",
              UN = "univ_comp", HE = "hedo_comp")
tenvalplot(df, valueMap)
}
